+++ test/reporter-cucumber-html/check-report/passed.test.ts
@@ -49,8 +49,8 @@ test('Scenario: Scenario with attachments', async ({ page }) => {
     'text attachmentsome text', // no space between 'attachment' and 'some'
     'image attachment inline',
     'image attachment as file',
-    'stdout123 some logs',
   ]);
+  await expect(scenario.getLogs()).toHaveText(['123 some logs']);
 });
 
 test('Scenario: Scenario with all keywords and success hooks', async ({ page }) => {